+++ FRAuth/FRAuth/Callbacks/CallbackFactory.swift
@@ -2,7 +2,7 @@
 //  CallbackFactory.swift
 //  FRAuth
 //
-//  Copyright (c) 2019-2021 ForgeRock. All rights reserved.
+//  Copyright (c) 2019-2022 ForgeRock. All rights reserved.
 //
 //  This software may be modified and distributed under the terms
 //  of the MIT license. See the LICENSE file for details.
@@ -38,6 +38,7 @@ import Foundation
          19. WebAuthnAuthenticationCallback
          20. IdPCallback
          21. SelectIdPCallback
+         22. DeviceBindingCallback
  */
 @objc(FRCallbackFactory)
 public class CallbackFactory: NSObject {
@@ -78,7 +79,8 @@ public class CallbackFactory: NSObject {
             CallbackType.WebAuthnAuthenticationCallback.rawValue: WebAuthnAuthenticationCallback.self,
             CallbackType.WebAuthnRegistrationCallback.rawValue: WebAuthnRegistrationCallback.self,
             CallbackType.IdPCallback.rawValue: IdPCallback.self,
-            CallbackType.SelectIdPCallback.rawValue: SelectIdPCallback.self
+            CallbackType.SelectIdPCallback.rawValue: SelectIdPCallback.self,
+            CallbackType.DeviceBindingCallback.rawValue: DeviceBindingCallback.self
         ]
     }